+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Moved fr.alchemy.utilities.pool package to fr.alchemy.utilities.collections.pool.
 - Added Version class to represent and compare sequence-based software versioning schemes.
 - Instantiator can now invoke a method using its name directly (without having the Method instance).
+- Tested Instantiator and Version class using appropriate test-cases.
 - Some JavaDoc polishing and rewriting as always.
  
 # 0.1.0 (Released March 27, 2020)